frontend: Extract VersionButton from Sidebar
